+++ src/server.rs
@@ -1,6 +1,7 @@
 pub mod client_entity_map;
 pub mod connected_clients;
 pub(super) mod despawn_buffer;
+pub mod events;
 pub(super) mod removal_buffer;
 pub(super) mod replicated_archetypes;
 pub(super) mod replication_messages;
@@ -23,11 +24,11 @@ use bevy::{
 };
 
 use crate::core::{
+    channels::{ReplicationChannel, RepliconChannels},
     common_conditions::{server_just_stopped, server_running},
     ctx::SerializeCtx,
-    replication_fns::ReplicationFns,
+    replication_registry::ReplicationRegistry,
     replication_rules::ReplicationRules,
-    replicon_channels::{ReplicationChannel, RepliconChannels},
     replicon_tick::RepliconTick,
     ClientId,
 };
@@ -65,6 +66,9 @@ impl Default for ServerPlugin {
     }
 }
 
+/// Server functionality and replication sending.
+///
+/// Can be disabled for client-only apps.
 impl Plugin for ServerPlugin {
     fn build(&self, app: &mut App) {
         app.add_plugins((DespawnBufferPlugin, RemovalBufferPlugin))
@@ -225,7 +229,7 @@ impl ServerPlugin {
             ResMut<ClientBuffers>,
             ResMut<RepliconServer>,
         )>,
-        replication_fns: Res<ReplicationFns>,
+        registry: Res<ReplicationRegistry>,
         rules: Res<ReplicationRules>,
         server_tick: Res<ServerTick>,
         time: Res<Time>,
@@ -241,7 +245,7 @@ impl ServerPlugin {
         collect_changes(
             &mut messages,
             &replicated_archetypes,
-            &replication_fns,
+            &registry,
             &entities_with_removals,
             set.p0(),
             &change_tick,
@@ -303,7 +307,7 @@ fn collect_mappings(
 fn collect_changes(
     messages: &mut ReplicationMessages,
     replicated_archetypes: &ReplicatedArchetypes,
-    replication_fns: &ReplicationFns,
+    registry: &ReplicationRegistry,
     entities_with_removals: &EntityHashSet,
     world: &World,
     change_tick: &SystemChangeTick,
@@ -365,7 +369,7 @@ fn collect_changes(
                     )
                 };
 
-                let (component_fns, rule_fns) = replication_fns.get(replicated_component.fns_id);
+                let (component_fns, rule_fns) = registry.get(replicated_component.fns_id);
                 let ctx = SerializeCtx { server_tick };
                 let mut shared_bytes = None;
                 for (init_message, update_message, client) in messages.iter_mut_with_clients() {